fix: Move event utils class into deprecated package [DHIS2-1604]
